requirements.txt
altair==4.0.1
atomicwrites==1.3.0
attrs==19.3.0
autograd==1.3
backports.tempfile==1.0
backports.weakref==1.0.post1
blis==0.4.0
bokeh==1.4.0
Bottleneck==1.3.1
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chainer==6.5.0
chardet==3.0.4
chart-studio==1.0.0
cloudpickle==1.2.2
cmake==3.12.0
contextlib2==0.5.5
cufflinks==0.17.0
cupy-cuda101==6.5.0
cvxopt==1.2.4
cvxpy==1.0.25
cycler==0.10.0
dask==2.9.2
decorator==4.3.0
descartes==1.1.0
dill==0.3.1.1
distributed==1.25.3
dlib==19.18.0
dm-sonnet==1.35
docopt==0.6.2
docutils==0.15.2
dopamine-rl==1.0.5
easydict==1.9
ecos==2.0.7.post1
editdistance==0.5.3
fancyimpute==0.4.3
fastai==1.0.60
fastprogress==0.2.2
featuretools==0.4.1
filelock==3.0.12
fsspec==0.6.2
future==0.16.0
gensim==3.6.0
gin-config==0.3.0
glob2==0.7
graph-nets==1.0.5
graphviz==0.10.1
greenlet==0.4.15
gym==0.15.6
HeapDict==1.0.1
humanize==0.5.1
hyperopt==0.1.2
ideep4py==2.0.0.post3
image==1.5.28
imageio==2.4.1
imagesize==1.2.0
imbalanced-learn==0.4.3
imblearn==0.0
imgaug==0.2.9
importlib-metadata==1.5.0
imutils==0.5.3
intel-openmp==2020.0.133
intervaltree==2.1.0
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
ipython-sql==0.3.9
ipywidgets==7.5.1
jax==0.1.58
jaxlib==0.1.38
Jinja2==2.10.1
jpeg4py==0.1.4
kapre==0.1.3.1
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-vis==0.4.1
kfac==0.2.0
kiwisolver==1.1.0
knnimpute==0.1.0
lightgbm==2.2.3
llvmlite==0.31.0
lucid==0.3.8
matplotlib==3.1.3
matplotlib-venn==0.11.5
mesh-tensorflow==0.1.9
missingno==0.4.2
mizani==0.6.0
mkl==2019.0
mlxtend==0.14.0
more-itertools==8.2.0
moviepy==0.2.3.5
mpi4py==3.0.3
mpmath==1.1.0
multiprocess==0.70.9
multitasking==0.0.9
natsort==5.5.0
nibabel==2.3.3
np-utils==0.5.12.1
numba==0.47.0
numexpr==2.7.1
numpy==1.17.5
nvidia-ml-py3==7.352.0
okgrade==0.4.3
opencv-contrib-python==4.1.2.30
opencv-python==4.1.2.30
opt-einsum==3.1.0
osqp==0.6.1
packaging==20.1
pandas==0.25.3
pandas-datareader==0.7.4
pandas-gbq==0.11.0
pandas-profiling==1.4.1
pandocfilters==1.4.2
parso==0.6.1
pathlib==1.0.1
patsy==0.5.1
pickleshare==0.7.5
Pillow==6.2.2
pip-tools==4.2.0
plac==0.9.6
plotly==4.4.1
prefetch-generator==1.0.1
protobuf==3.10.0
psutil==5.4.8
ptyprocess==0.6.0
py==1.8.1
pycparser==2.19
pydot==1.3.0
pydot-ng==2.0.0
pydotplus==2.0.2
pymc3==3.7
PyOpenGL==3.1.5
pypng==0.0.20
pyrsistent==0.15.7
pystan==2.19.1.1
pytest==3.6.4
python-dateutil==2.6.1
python-louvain==0.13
python-slugify==4.0.0
python-utils==2.3.0
PyWavelets==1.1.1
PyYAML==3.13
regex==2019.12.20
requests==2.19.0
resampy==0.2.2
retrying==1.3.3
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.1.0
scs==2.1.1.post2
seaborn==0.10.0
Send2Trash==1.5.0
setuptools-git==1.2
Shapely==1.7.0
simplegeneric==0.8.1
six==1.13.0
sklearn==0.0
sklearn-pandas==1.8.0
sortedcontainers==2.1.0
srsly==1.0.1
stable-baselines==2.2.1
statsmodels==0.10.2
sympy==1.1.1
tblib==1.6.0
tensor2tensor==1.14.1
tensorboard==1.15.0
tensorboardX==2.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gan==2.0.0
tensorflow-hub==0.7.0
tensorflow-metadata==0.21.1
tensorflow-privacy==0.2.2
tensorflow-probability==0.7.0
testpath==0.4.4
text-unidecode==1.3
tflearn==0.3.2
Theano==1.0.4
toolz==0.10.0
torch==1.4.0+cu100
torchsummary==1.5.1
torchtext==0.3.1
torchvision==0.5.0+cu100
traitlets==4.3.3
umap-learn==0.3.10
xarray==0.14.1
xgboost==0.90
yellowbrick==0.9.1
zict==1.0.0
zipp==2.1.0
